show reviews in product detal

+++ products/models.py
@@ -69,7 +69,7 @@ class Review(models.Model):
     user = models.ForeignKey(User,verbose_name=('user'),related_name='review_author',on_delete=models.CASCADE)
     Product = models.ForeignKey(Product,verbose_name=('product'),related_name='product_review',on_delete=models.CASCADE)
     review =models.TextField(_('review'),max_length=500)
-    rate =models.BigIntegerField(_('rate'))
+    rate =models.IntegerField(_('rate'))
     create_date =models.DateTimeField(_('creat_date'),default=timezone.now)
 
     def __str__(self):
